Combine mesh-surface related tests into same file
